+++ CHANGES.rst
@@ -100,8 +100,8 @@ Bug Fixes
 
 - Fixed API hint styling for viewer labels. [#4003]
 
-- Fixed bug when opening plot options from shortcut button when jdaviz and lcviz are simultaneously
-  imported [#4027]
+- Fixed bug when opening plot options or export from shortcut button when jdaviz
+  and lcviz are simultaneously imported. [#4027]
 
 Cubeviz
 ^^^^^^^